LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-largeop-004-expected.txt


PASS Operator dictionary chunk 1 - largeop
PASS Operator dictionary chunk 2 - largeop
FAIL Operator dictionary chunk 3 - largeop assert_approx_equals: Largeop property for ⨝ prefix should be 'true' expected 100 +/- 1 but got 50
PASS Operator dictionary chunk 4 - largeop
largeop for "⨝" (prefix):
⨝
 VS
⨝
